Refactor advice and helper in the freemarker instrumentation
