+++ dd-java-agent/instrumentation/freemarker-2.3.24/src/main/java/datadog/trace/instrumentation/freemarker24/DollarVariableDatadogAdvice.java
@@ -1,24 +1,25 @@
-package freemarker.core;
+package datadog.trace.instrumentation.freemarker24;
 
 import datadog.trace.api.iast.InstrumentationBridge;
 import datadog.trace.api.iast.Sink;
 import datadog.trace.api.iast.VulnerabilityTypes;
 import datadog.trace.api.iast.sink.XssModule;
-import freemarker.template.TemplateException;
+import freemarker.core.DollarVariable24Helper;
+import freemarker.core.Environment;
 import net.bytebuddy.asm.Advice;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public final class DollarVariable24DatadogAdvice {
+public final class DollarVariableDatadogAdvice {
 
-  protected static final Logger log = LoggerFactory.getLogger(DollarVariable24DatadogAdvice.class);
+  protected static final Logger log = LoggerFactory.getLogger(DollarVariableDatadogAdvice.class);
 
   public static class DollarVariableAdvice {
 
     @Advice.OnMethodEnter(suppress = Throwable.class)
     @Sink(VulnerabilityTypes.XSS)
     public static void onEnter(
-        @Advice.Argument(0) final Environment environment, @Advice.This final DollarVariable self) {
+        @Advice.Argument(0) final Environment environment, @Advice.This final Object self) {
       if (environment == null || self == null) {
         return;
       }
@@ -29,15 +30,9 @@ public final class DollarVariable24DatadogAdvice {
       if (DollarVariable24Helper.fetchAutoEscape(self)) {
         return;
       }
-      String charSec = null;
-      try {
-        charSec = (String) self.calculateInterpolatedStringOrMarkup(environment);
-      } catch (TemplateException e) {
-        log.debug("Failed to get DollarVariable templateModel", e);
-        return;
-      }
+      String charSec = DollarVariable24Helper.fetchCharSec(self, environment);
       final String templateName = environment.getMainTemplate().getName();
-      final int line = self.beginLine;
+      final int line = DollarVariable24Helper.fetchBeginLine(self);
       xssModule.onXss(charSec, templateName, line);
     }
   }
